requirements.txt
fastapi==0.109.0
uvicorn==0.22.1
pandas==2.1.0
scikit-learn==1.3.0
scikeras==0.11.0
tensorflow-cpu==2.14.0
snowflake-connector-python==3.1.3
boto3==1.28.45
